Add function to process AssetSearchFormat configuration
The AssetSearchFormat configuration option for assets accepts
both a single string defining a search results format and a
hashref that defines a different format per catalog. This was
needed for the initial asset search implementation because
there was no other way to modify the search results format. Add a
function to process both formats in AssetSQL.

AssetSQL now provides a user interface to modify the format
on the Advanced tab, so for the hashref version define a single
default search result format key called AssetSearchFormat.